+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 import useRecommendData from './useRecommendData';
+import useSearchData from './useSearchData';
 import useAuth from './useAuth';
 
-export { useAuth, useRecommendData };
+export { useAuth, useRecommendData, useSearchData };
